#assistants: fixes for MSVC

+++ toonz/sources/tnztools/fullcolorbrushtool.cpp
@@ -591,7 +591,7 @@ void FullColorBrushTool::loadPreset() {
   try  // Don't bother with RangeErrors
   {
     m_thickness.setValue(
-        TIntPairProperty::Value(std::max((int)preset.m_min, 1), preset.m_max));
+        TIntPairProperty::Value(std::max((int)preset.m_min, 1), (int)preset.m_max));
     m_hardness.setValue(preset.m_hardness, true);
     m_opacity.setValue(
         TDoublePairProperty::Value(preset.m_opacityMin, preset.m_opacityMax));
